+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **Date**: Thu Oct 25 11:07:51 CEST 2018
 
 * [CHANGE] Add user namespace to localStorage #101
+* [FIX] n.data[b].client is undefined after open events #102
 
 
 ## 3.0.4